+++ src/collective/mustread/interfaces.py
@@ -72,9 +72,7 @@ class ITracker(Interface):
     - mark_read(obj, 'maryjane')
     - has_read(obj, 'johndoe')
     - who_read(obj)
-
-    This is the initial implementation. It is not required (and at this stage
-    not even possible) to schedule a ``must_read(obj)`` to track reads.
+    - uids_read('johndoe')
 
     Scheduled for later extension:
 
@@ -82,12 +80,14 @@ class ITracker(Interface):
        track reads on that object, and query which users have failed to
        meet the requirment of reading the object in time:
 
-    - must_read(obj, ['johndoe', 'maryjane'], nextweek)
+    - schedule_must_read(obj, ['johndoe', 'maryjane'], nextweek)
     - mark_read(obj, 'johndoe')
-    - who_unread(obj)
+    - who_must_read(obj)
+    - what_to_read(context)
+    - what_to_read(userid)
 
     The second scenario is a superset of the first. Specifically, also
-    users who were not scheduled as ``must_read`` will be tracked and
+    users who were not scheduled as ``schedule_must_read`` will be tracked and
     returned when you query ``who_read(obj)``.
 
     Note that we use userids not usernames, since userids will be more stable.
@@ -165,36 +165,38 @@ class ITracker(Interface):
         :rtype: Generator
         '''
 
-    #  ------- @frisi --- TO BE IMPLEMENTED ---------------------------------
-
-    def schedule_must_read(obj, userids, deadline=None):
-        '''
-        Schedule that an object must be read by some users before a deadline.
+    def schedule_must_read(obj, userids, deadline):
+        '''Schedule that an object must be read by some users before a deadline.
 
         Calling this method is optional. An object does not have to be
         scheduled as 'must-read' in order to track reads.
 
         Calling this method and scheduling an object as 'must-read' enables
         tracking of 'unread' status for specific sets of users and querying
-        for those via the ``who_unread`` method.
+        for those via the ``who_must_read`` method.
 
         :param obj: Object that should be read by users
         :type obj: Content object (must be IUUID resolvable)
         :param userids: Userids of the users that should view the object.
-        :type userid: List
+        :type userids: List
         :param deadline: Deadline before which users should view the object.
         :type deadline: datetime
+        :returns: List of userids a mustread deadline has been scheduled for
+                  (existing deadlines do not get changed)
+        :rtype: list
         '''
 
-    def who_unread(obj, force_deadline=True):
-        '''
-        For an object scheduled as must-read, query which users have not
-        read the object. Only makes sense of you first scheduled ``must_read``
-        which is optional.
+    def who_must_read(obj):
+        # def who_must_read(obj, force_deadline=True):
+        '''For an object scheduled as must-read, query which users have not
+        read the object.
 
-        This only queries for users who have been  explicitly scheduled via
-        the ``must_read()`` method. If no users were scheduled, returns an
-        empty list.
+        Only makes sense of you first scheduled a must read
+        via ``schedule_must_read`` - which is optional.
+
+        This only queries for users who have been explicitly scheduled via
+        the ``schedule_must_read()`` method. If no users were scheduled,
+        returns an empty dictionary.
 
         If ``force_deadline`` is True, returns scheduled users who
         did not read the object before the deadline. This includes users
@@ -203,10 +205,91 @@ class ITracker(Interface):
         If ``force_deadline`` is False, returns scheduled users who
         did not read the object at all.
 
+        @guido: i'd suggest to remove force_deadline as we have the get_stats
+        method to get a detailed report which users read and when
+
         :param obj: Object that should be read by users
         :type obj: Content object (must be IUUID resolvable)
         :param force_deadline: Whether to ignore reads after the deadline
         :type force_deadline: Bool
-        :returns: Userids of users scheduled for reading which did not read
-        :rtype: List
+        :returns: Dictionary with userids of users scheduled for reading which
+                  did not read and their deadline (``{userid: deadline}``)
+        :rtype: dict
+        '''
+
+    def what_to_read(context=None, userid=None):
+        '''Query which content items have open read requests.
+
+        This only queries for objects who have been explicitly scheduled via
+        the ``schedule_must_read()`` method.
+        If no read requests have been scheduled an empty list is returned.
+
+        If context is given, limit the result to objects located within
+        this context. If context is ``None`` search within the plonesite.
+
+        If userid is given, limit results to objects with an open read request
+        for this userid.
+
+        @guido: if we implement the force_deadline here, we'd have an easy way
+        to list all items with missed deadlines too. alternatively we can make
+        get_report more powerful/complex
+
+        :param context: Context to search for objects with open read requests
+        :type context: Content object (must be IUUID resolvable)
+        :param userid: Userid to search for open read requests for.
+        :type userid: string
+        :returns: List of content objects with open read requests
+        :rtype: Generator
+        '''
+        # xxx guido: generator or list? (we intend to use all elements anyway)
+
+    def get_report(context=None, include_children=True, userid=None,
+                   start_date=None):
+        '''Return all mustread database entries for context.
+        (useful for exporting data or displaying detailed reports)
+
+        If context is None, defaults to plone site.
+
+        If include_children is ``True`` (which is the default), include
+        mustread records for child-objects of context.
+
+        If userid is given, only return entries for this user
+
+        If start_date is given, only return entries created after start_date
+        (``scheduled_at`` or ``read_at`` >= ``start_date``
+
+        :param context: Context to create mustread report on
+        :type context: Content object (must be IUUID resolvable)
+        :param include_children: Whether to include child-objects of context
+        :type include_children: Bool
+        :param userid: only report records for specified user
+        :type userid: string
+        :param start_date: Filter out entries older than start_date
+        :type read_at: datetime
+        :returns: List of dictionaries, each representing a mustread record
+                  ``[{userid, uid, status, read_at, deadline, scheduled_at,
+                      type, path, ...}]``
+        :rtype: Generator
+        '''
+        # xxx guido: generator or list? (we intend to use all elements anyway)
+
+    def unschedule_must_read(obj=None, userids=None):
+        '''Maintenance method to remove open must-read requests for an object
+        and/or a specified userid
+
+        @guido decide if we need/want this
+        '''
+
+    def accept_not_read(obj=None, userids=None):
+        '''Maintenance method to mark open must-read requests for an object
+        and/or a specified userid with status 'not-read' and leaving the
+        read_at date empty.
+        this way they object is not listed as what_to_read
+        and the user ist not listed in who_must_read
+
+        this would allow admins to clear certain objects form the must-read
+        list or accept that a user (that might be ill) will not be required to
+        read a content item
+
+        @guido decide if we need/want this
         '''