+++ CHANGES.rst
@@ -8,6 +8,12 @@ Changelog
 - Added the possibility to specify engine parameters through the registry
   [ale-rt]
 
+- Implemented API for scheduling items as must-read for certain users.
+  (see collective.contentrules.mustread for usage)
+
+  This required new database columns. The provided upgrade step works for sqlite databases
+  but might need changes for mysql or postgres.
+
 - Allow to create and configure a database file by installing the `database` setup profile